+++ plugins/main/public/components/add-modules-data/WzSampleDataWrapper.js
@@ -29,7 +29,6 @@ import {
   withReduxProvider,
 } from '../../components/common/hocs';
 import { compose } from 'redux';
-import { WAZUH_ROLE_ADMINISTRATOR_NAME } from '../../../common/constants';
 
 export class WzSampleDataProvider extends Component {
   constructor(props) {
@@ -68,5 +67,5 @@ export class WzSampleDataProvider extends Component {
 export const WzSampleDataWrapper = compose(
   withErrorBoundary,
   withReduxProvider,
-  withUserAuthorizationPrompt(null, [WAZUH_ROLE_ADMINISTRATOR_NAME]),
+  withUserAuthorizationPrompt(null, { isAdmininistrator: true }),
 )(WzSampleDataProvider);