+++ plugins/main/public/controllers/management/monitoring.js
@@ -18,7 +18,6 @@ import { TabVisualizations } from '../../factories/tab-visualizations';
 import store from '../../redux/store';
 import { updateGlobalBreadcrumb } from '../../redux/actions/globalBreadcrumbActions';
 import { ModulesHelper } from '../../components/common/modules/modules-helper';
-import { WAZUH_ROLE_ADMINISTRATOR_NAME } from '../../../common/constants';
 import { getCore, getDataPlugin } from '../../kibana-services';
 import { cluster, endpointSummary } from '../../utils/applications';